refactor: set main and source in package.json
