+++ package.json
@@ -5,6 +5,8 @@
   "engines": {
     "node": ">=14.0.0"
   },
+  "main": "dist/index.js",
+  "source": "src/index.ts",
   "description": "Graasp Translations",
   "repository": {
     "type": "git",
@@ -45,6 +47,6 @@
   },
   "packageManager": "yarn@3.2.0",
   "dependencies": {
-    "i18n": "0.14.1"
+    "i18next": "21.6.11"
   }
 }
