+++ tsconfig.json
@@ -2,7 +2,7 @@
   "compilerOptions": {
     "target": "es6",
     "module": "commonjs",
-    // "strict": true,
+    "strict": true,
     "outDir": "dist",
     "esModuleInterop": true,
     "skipLibCheck": true,
